Created User profile
Now the user can click on other members of common events, challenges. It navigates to the clicked user's profile page. There we can see their profile picture, and description, what was the last challenge of the member. We can also send the user messages by clicking a button. We can delete our own profile, or admins can do it.